Testcase and fix for issue 1348
